pyproject.toml
[project]
name = "medhelper"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-factoryboy (>=2.7.0,<3.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
]

[tool.poetry]
name = "userhub"
version = "0.1.0"
description = ""
authors = ["takelele <work.mnikulin@gmail.com>"]
readme = "README.md"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
